+++ CHANGELOG.md
@@ -15,6 +15,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   ``APIClient.bocceclusters`` -> ``APIClient.bocce_clusters``
   ``APIClient.matchtargets`` -> ``APIClient.match_targets``
   ``APIClient.remotehosts`` -> ``APIClient.remote_hosts``
+- ``civis.io.csv_to_civis`` and ``civis.io.dataframe_to_civis`` functions now use
+  ``civis.io.file_to_civis`` and ``civis.io.civis_file_to_table`` functions instead
+  of separate logic
+- ``civis.io.file_to_civis``, ``civis.io.csv_to_civis`` and ``civis.io.dataframe_to_civis``
+  now support files over 5GB
 
 ### Fixed
 - Fixed parsing of multiword endpoints. Parsing no longer removes underscores
@@ -29,6 +34,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   and ``civis.parallel.infer_backend_factory`` in order to set the joblib
   backend in the container for nested calls to ``joblib.Parallel``.
 - Added the PyPI trove classifiers for Python 3.4 and 3.6 (#152).
+- ``civis.io.civis_file_to_table`` function to import an existing Civis file
+  to a table
+- ``civis.io.file_to_civis`` function will now automatically retry uploads to
+  the Civis Platform up to 5 times if is there is an HTTPError, ConnectionError
+  or ConnectionTimeout
+
+### Performance Enhancements
+- ``civis.io.file_to_civis`` now takes advantage of multipart uploads to chunk
+  files and perform I/O in parallel
 
 ## 1.6.2 - 2017-09-08
 ### Changed